refactor project to change structure
