+++ .github/workflows/deploy.yml
@@ -25,8 +25,17 @@ jobs:
         uses: docker/build-push-action@v2
         with:
           context: .
+          file: ./docker/telebot/Dockerfile
           push: true
           tags: anthonyfong/leetcode-tele-bot:latest
+      - name: Build and Push to Docker Registry
+        uses: docker/build-push-action@v2
+        with:
+          context: .
+          file: ./docker/leetcode_server/Dockerfile
+          push: true
+          tags: anthonyfong/leetcode-server:latest
+
   deploy:
     needs: build
     runs-on: ubuntu-latest
