--- Pipfile
@@ -1,25 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-black = "*"
-isort = "*"
-pre-commit = "*"
-flake8 = "*"
-pytest = "*"
-pytest-cov = "*"
-
-[packages]
-python-dotenv = "*"
-pytelegrambotapi = "*"
-fastapi = "*"
-
-[requires]
-python_version = "3.7"
-
-[pipenv]
-allow_prereleases = true
-
-[scripts]
