--- leetcode_tele_bot/telebot_service/handlers/__init__.py
@@ -1,2 +1,0 @@
-import leetcode_tele_bot.telebot_service.handlers.leetcode as leetcode
-import leetcode_tele_bot.telebot_service.handlers.telebot as telebot
