--- leetcode_tele_bot/telebot_service/handlers/telebot.py
@@ -1,5 +1,0 @@
-from leetcode_tele_bot.telebot_service.api.telegram import TelegramApiClient
-
-
-def send_welcome(message):
-    TelegramApiClient.reply_to(message, "Howdy, how are you doing?")
