+++ libs/common/common/text.py
