+++ libs/common/pyproject.toml
@@ -1,0 +1,17 @@
+[tool.poetry]
+name = "common"
+version = "0.1.0"
+description = ""
+authors = ["Anthony <anthonyfong100@gmail.com>"]
+
+[tool.poetry.dependencies]
+python = "^3.7"
+
+
+[tool.poetry.dev-dependencies]
+pytest = "^5.2"
+pytest-cov = "^2.12.1"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
