+++ services/telebot_backend_svc/pyproject.toml
@@ -1,0 +1,19 @@
+[tool.poetry]
+name = "telebot_backend_svc"
+version = "0.1.0"
+description = ""
+authors = ["Anthony <anthonyfong100@gmail.com>"]
+
+[tool.poetry.dependencies]
+python = "^3.7"
+fastapi = "^0.65.1"
+uvicorn = "^0.14.0"
+common = ""
+
+[tool.poetry.dev-dependencies]
+pytest = "^5.2"
+common = {path = "../../libs/common", develop = true}
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
