+++ services/telebot_backend_svc/telebot_backend_svc/__init__.py
@@ -1,0 +1,1 @@
+__version__ = "0.1.0"
