+++ services/telebot_backend_svc/tests/test_telebot_backend_svc.py
@@ -1,0 +1,5 @@
+from telebot_backend_svc import __version__
+
+
+def test_version():
+    assert __version__ == "0.1.0"
