+++ services/telegram_svc/Dockerfile
@@ -1,0 +1,9 @@
+FROM python:slim
+
+COPY ./dist/deps/*.whl ./
+RUN pip install *.whl && rm -rf *.whl
+
+COPY ./dist/*.whl ./
+RUN pip install *.whl && rm -rf *.whl
+
+CMD ["python", "-m", "telegram_svc.main"]
