+++ services/telegram_svc/Makefile
@@ -1,0 +1,4 @@
+_all: all
+_clean: clean 
+
+include ../../tools/Packages.mk
