+++ services/telegram_svc/pyproject.toml
@@ -1,0 +1,20 @@
+[tool.poetry]
+name = "telegram_svc"
+version = "0.1.0"
+description = ""
+authors = ["Anthony <anthonyfong100@gmail.com>"]
+
+[tool.poetry.dependencies]
+python = "^3.7"
+pyTelegramBotAPI = "^3.7.9"
+python-dotenv = "^0.17.1"
+common = ""
+
+[tool.poetry.dev-dependencies]
+pytest = "^5.2"
+common = {path = "../../libs/common", develop = true}
+
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
