+++ services/telegram_svc/telegram_svc/handlers/__init__.py
@@ -1,0 +1,2 @@
+import telegram_svc.handlers.leetcode as leetcode
+import telegram_svc.handlers.telebot as telebot
