+++ services/telegram_svc/telegram_svc/handlers/telebot.py
@@ -1,0 +1,5 @@
+from telegram_svc.api.telegram import TelegramApiClient
+
+
+def send_welcome(message):
+    TelegramApiClient.reply_to(message, "Howdy, how are you doing?")
