+++ services/telegram_svc/tests/test_telegram_svc.py
@@ -1,0 +1,5 @@
+from telegram_svc import __version__
+
+
+def test_version():
+    assert __version__ == "0.1.0"
